first

+++ src/main/java/com/aston/productservice/service/ProductServiceImpl.java
@@ -1,6 +1,7 @@
 package com.aston.productservice.service;
 
 import com.aston.productservice.dto.ProductDTO;
+import com.aston.productservice.exception.ProductNotFoundException;
 import com.aston.productservice.mapper.ProductMapper;
 import com.aston.productservice.model.Product;
 import com.aston.productservice.repository.ProductRepository;
@@ -45,19 +46,35 @@ public class ProductServiceImpl implements ProductService {
         return productMapper.toDto(product);
     }
 
+
     @Override
     public Optional<ProductDTO> updateProduct(Long id, ProductDTO productDto) {
         return productRepository.findById(id)
                 .map(product -> {
-                    product.setProductName(productDto.getProductName());
-                    product.setProductType(productDto.getProductType());
-                    product.setProductPrice(productDto.getProductPrice());
+                    product.setName(productDto.getProductName());
+                    product.setType(productDto.getProductType());
+                    product.setPrice(productDto.getProductPrice());
                     return productMapper.toDto(productRepository.save(product));
                 });
     }
 
     @Override
-    public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
+    public boolean deleteProduct(Long id) {
+        boolean exists = productRepository.existsById(id);
+        if (exists) {
+            productRepository.deleteById(id);
+            return true;
+        } else {
+            throw new ProductNotFoundException("Product with ID " + id + " not found.");
+        }
+    }
+    @Override
+    public List<ProductDTO> findProductsByNameContaining(String name) {
+        // Использование метода репозитория для поиска продуктов по содержанию имени, игнорируя регистр
+        List<Product> products = productRepository.findByNameContainingIgnoreCase(name);
+        // Преобразование списка продуктов в список DTO
+        return products.stream()
+                .map(productMapper::toDto)
+                .collect(Collectors.toList());
     }
 }
